Adds socket pre-connect test.
Jira: OTPL-2001

Previously, we'd attempt to connect just with the Postgres library.
However, it unconditionally performs some annoying logging if it fails
to connect, and it's natural to fail to connect a bit when starting up.
Thus, we add a simple socket connection attempt of our own before trying
to execute a Postgres connection.