Make Darwin bundle fully self-contained by including libpq
libpq is the C Postgres client library. It is used by bin/initdb which is
packaged and run by this project.

While the Linux and Windows tarballs contain libpq, the Mac (Darwin) one does
not. This mostly works out ok because Macs ship with /usr/lib/libpq.5.dylib (and
you can install your own libpq to other places like /usr/local/lib with, eg,
Homebrew), but this might be a surprising version or may be misinstalled. (I
discovered this problem when my Homebrew Postgres installation was misinstalled
and it broke EmbeddedPostgres.)

By packaging libpq with the Darwin bundle, EmbeddedPostgres can be fully
selfcontained on Mac.

You can verify this as follows:

    ./repack-postgres.sh
    mkdir unpacked
    cd unpacked
    tar xJf ../target/generated-resources/postgresql-Darwin-x86_64.txz
    DYLD_PRINT_LIBRARIES=YES ./bin/initdb --version 2>&1 | grep libpq

Before this change, the final command reads:

    dyld: loaded: /usr/lib/libpq.5.dylib

After this change, it reads:

    dyld: loaded: /private/tmp/otj-pg-embedded/unpacked/./bin/../lib/libpq.5.dylib